backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
openai>=1.0.0
chromadb==0.4.18
langchain==0.1.20
langchain-community==0.0.38
python-multipart==0.0.6
python-dotenv==1.0.0
huggingface-hub==0.20.3
SpeechRecognition==3.10.0
sentence-transformers==2.2.2
edge-tts==6.1.9